requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
pydantic
passlib
bcrypt
python-jose
